move kitti to tests
